refactor: block 컴포넌트의 handleKeyDown 함수 파일로 분리
